+++ RepoAdminMenu.cs
@@ -13,7 +13,7 @@ namespace RepoAdminMenu {
 
         private const string mod_guid = "proferabg.REPO.RepoAdminMenu";
         private const string mod_name = "Repo Admin Menu";
-        private const string mod_version = "1.0.4";
+        private const string mod_version = "1.0.5";
 
         private static RepoAdminMenu _plugin;